requirements.txt
torch==1.11.0
torchvision==0.12.0
opencv-python==4.5.5.64
numpy==1.22.3
Pillow==9.0.1
PyQt5
open3d